backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy>=1.26.0,<2.0.0
Pillow==10.1.0
requests==2.31.0
opencv-python==4.8.1.78
python-dotenv==1.0.0
#gunicorn==21.2.0
pymongo==4.6.1
bcrypt==4.1.2
Flask-JWT-Extended==4.6.0
Flask-Session==0.4.0
google-genai
# For actual ML model deployment (uncomment when ready)
tensorflow>=2.15.0
torch==2.2.2
torchvision==0.17.2
transformers==4.36.0
